Lab13

+++ src/main/java/com/example/songr/AlbumRepository.java
@@ -2,4 +2,4 @@ package com.example.songr;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface AlbumRepository extends CrudRepository<Album,Integer> {}
+public interface AlbumRepository extends CrudRepository<Album,Long> {}
