+++ src/main/java/com/example/songr/songRepository.java
@@ -2,4 +2,4 @@ package com.example.songr;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface songRepository extends CrudRepository<Song,Integer> {}
+public interface songRepository extends CrudRepository<Song,Long> {}
